refactor: rename track_factory to track_lookup - inline with the exported function name
